% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_existing_model}
\alias{get_existing_model}
\title{Get Existing Model}
\usage{
get_existing_model(model_path)
}
\arguments{
\item{model_path}{}
}
\value{

}
\description{
List all models and downloads last one, assumes order by name import
dowloaded model
}
