% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_score}
\alias{get_score}
\title{Get Score}
\usage{
get_score(params_dict, model_package)
}
\arguments{
\item{model_package}{}
}
\value{

}
\description{
Receiva all features in a dic, Predicts score and return it
}
