% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{read_input_data}
\alias{read_input_data}
\title{Read Input Data}
\usage{
read_input_data(
  data_path = "data",
  data_file = "auto.xlsx",
  container_url = "https://storageaccountcsapp9d08.blob.core.windows.net/",
  key = Sys.getenv("STORAGE_ACCOUNT_KEY")
)
}
\arguments{
\item{data_path}{(string) path to data}

\item{data_file}{(string) a file name with .xlsx extension}

\item{container_url}{(string) azure container url}

\item{key}{(string) Auth key from Azure Blob Storage}
}
\value{
a data.frame
}
\description{
Downloads specified data file from specified container,
import this data to a data.frame an returns it.
Data file must be Excel file, container URL and key
must be from Azure Blob Storage.
}
