% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{run_train}
\alias{run_train}
\title{Run Train}
\usage{
run_train(
  data_path,
  data_file,
  model_path,
  model_name,
  container_url,
  key,
  seed = 1992
)
}
\arguments{
\item{key}{}
}
\value{

}
\description{
Run the complete model training flow Downloads and imports data from cloud,
run all transformation pipeline, fit model and save it to cloud.
}
