% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{save_model}
\alias{save_model}
\title{Save Model}
\usage{
save_model(
  model,
  model_path = "models/",
 
    model_name = glue::glue("auto_model_{strftime(start_time, \\"\%Y-\%m-\%d--\%H-\%M-\%S\\")}.rds"),
  container_url = "https://storageaccountcsapp9d08.blob.core.windows.net/",
  key = Sys.getenv("STORAGE_ACCOUNT_KEY")
)
}
\arguments{
\item{key}{}
}
\value{

}
\description{
First, serialize model object to pickle and save to disk, then uploads it to
cloud.
}
