% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{update_model}
\alias{update_model}
\title{Update Model}
\usage{
update_model(model_path, container_url, key)
}
\arguments{
\item{key}{}
}
\value{

}
\description{
List all models and downloads last one, assumes order by name import
dowloaded model
}
